+++ test/WalletSendEthTest.t.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.8.18;
 import {Test, console} from "forge-std/Test.sol";
 import {Wallet} from "../src/Wallet.sol";
 
-
 contract WalletSendEthTest is Test {
     Wallet public wallet;
 
@@ -13,16 +12,16 @@ contract WalletSendEthTest is Test {
     }
 
     function _send(uint256 amount) private {
-        (bool ok, ) = address(wallet).call{value: amount}("");
+        (bool ok,) = address(wallet).call{value: amount}("");
         require(ok, "send failed");
     }
 
     function testEthBalance() public {
-        console.log("Eth balance: ", address(this).balance/1e18);
+        console.log("Eth balance: ", address(this).balance / 1e18);
     }
 
     function testSendEth() public {
-        uint bal = address(wallet).balance;
+        uint256 bal = address(wallet).balance;
         //address(1)を100etherに設定
         deal(address(1), 100);
         assertEq(address(1).balance, 100);
@@ -40,8 +39,7 @@ contract WalletSendEthTest is Test {
         //上の３行をhoax()を使って短縮
         hoax(address(1), 456);
         _send(456);
-        
+
         assertEq(address(wallet).balance, bal + 123 + 456);
     }
-
 }
